feat: Added lora msg_id and button push to sync
This first change is a first step in adding retry logic to handle the case of a lora cmd being received but the receipt being lost in transmisson.

The second change makes it so that the PolyPlug only starts listening when the sync button is pressed. This saves power and reduces un-sync from rogue transmissions.